+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
 ## New features: 
 
+  1. improve progress bar for data frames 
+
 ## Backward incompatible:  
  
 ## Bug fixes: